Add Object.cli_property decorator
This decorator enables complex command-line interfaces by grouping
arguments that are relevant for all (or most) subcommands on the
class itself, rather than duplicating those arguments on each method.

Object.cli_property can be applied to any `property` in the class.
If default values should be set for those properties, they should
be set in the class's `__init__`. See the updated `cli_object.py`
example for usage.

Note that when passing arguments through the command line, the
CLI property arguments can be passed before or after the command.
Currently the command can appear either first, or after the
CLI property arguments, and possibly after the command's arguments,
as we do not check that the arguments coming before the command
are CLI property arguments (though this could be done).

Other changes in this commit:

- The args possibly passed to the `CLIInterfaceAPI.__call__` method
  are now passed onto the interface call, which passes them on to the
  `parser.parse_known_args` call which actually uses them.